Move chrome memory to its own file.  Changed the option passing strategy.
