+++ lib/memory.js
@@ -9,91 +9,8 @@ const q = require('q'),
   mkdirp = require('mkdirp'),
   path = require('path'),
   exec = require('child_process').exec,
+  chromeMemory = require('./chrome-memory'),
   pusage = require('pidusage');
-
-/**
- * A set of checks to use when searching for a running Chrome process
- *
- * @typedef {{includes:boolean, text:string}[]} ValidChromeArgs
- */
-let chromeArguments;
-
-/**
- * The default ValidChromeArgs
- *
- * @type {ValidChromeARgs}
- */
-const defaultChromeArguments = [
-  {includes: true, text: 'test-type=protractor-memory'},
-  {includes: false, text: 'extension-process'}
-];
-
-/**
- * Find all windows that are running via Protractor and calculate
- * the current memory usage for each process.  Returns an array
- * of process information (sorted with the most recently started
- * process first)
- *
- * @returns {Promise} Promise that, when resolved, gives an array of Chrome process information
- */
-function getChromeProcessMemory() {
-  const promise = protractor.promise.defer();
-  let foundProcesses, statCallbackCount;
-
-  exec('ps -Ao pid,etime,args', function(err, stdout, stderr) {
-    if (stderr) promise.reject(stderr);
-
-    // One process per line
-    foundProcesses = stdout.split('\n').reduce((array, line) => {
-      if (isValid(line)) {
-        const tokens = line.split(/[ ]+/);
-
-        // First element is the pid
-        array.push({pid: tokens[0], etime: tokens[1], args: tokens[2]});
-      }
-      return array;
-    }, []);
-
-    // Wait for the proper amount of statistics to come back
-    statCallbackCount = foundProcesses.length;
-
-    // Sort by start time (most recent first)
-    foundProcesses = foundProcesses.sort(p => p.etime).reverse();
-
-    // Get the memory usage for each process
-    foundProcesses.forEach(process => {
-      // Need a closure here to keep track of processes
-      getUsedMemory(process);
-    });
-  });
-
-  return promise;
-
-  // We only care about chrome processes that have the correct flags passed to them.
-  function isValid(args) {
-    var valid = args.includes('Google Chrome Helper');
-
-    if (valid) {
-      _.forEach(chromeArguments, arg => {
-        const includes = args.includes(arg.text);
-        return valid = arg.includes ? includes : !includes; // Break out if valid is false
-      });
-    }
-
-    return valid;
-  }
-
-  // For a given process, calculate the used memory
-  function getUsedMemory(process) {
-    pusage.stat(process.pid, (e, stat) => {
-      if (e) promise.reject(e);
-      process.memory = stat.memory;
-
-      // Resolve the promise if this is the last one.
-      if (!--statCallbackCount) promise.fulfill(foundProcesses);
-    });
-  }
-};
 
 function runScenario(iterations, fn, options) {
   options = options || {};
@@ -104,11 +21,9 @@ function runScenario(iterations, fn, options) {
     progressBar: true,
     dir: 'results',
     name: 'memory-test',
-    chromeProcessArgs: defaultChromeArguments,
+    chromeProcessArgs: chromeMemory.defaultChromeArguments,
     units: 'MB'
   });
-
-  chromeArguments = options.chromeProcessArgs;
 
   var time = new Date().toISOString().replace(/\W/gi, '-');
   mkdirp.sync(options.dir);
@@ -145,14 +60,14 @@ function execute(iterations, fn, options) {
     });
   }
 
-  return recordMemory(measurements, options.units, options.finalPostGcSleep)
+  return recordMemory(measurements, options)
     .then(() => measurements);
 }
 
-function recordMemory(measurements, units, waitTime) {
+function recordMemory(measurements, options) {
   return garbageCollect()
-    .then(() => waitTime ? browser.sleep(waitTime) : undefined)
-    .then(() => getMemory(units))
+    .then(() => options.waitTime ? browser.sleep(options.waitTime) : undefined)
+    .then(() => getMemory(options))
     .then(memory => {
       measurements.push(memory);
       return memory;
@@ -166,8 +81,8 @@ function garbageCollect() {
 /**
  * Gets js heap total and used as well as chrome process memory
  */
-function getMemory(units) {
-  return q.all([getJavascriptMemory(units), getChromeMemory(units)])
+function getMemory(options) {
+  return q.all([getJavascriptMemory(options.units), getChromeMemory(options)])
     .then(results => {
       return _.extend({}, results[0], results[1]);
     });
@@ -183,8 +98,8 @@ function getJavascriptMemory(units) {
     });
 }
 
-function getChromeMemory(units) {
-  return getChromeProcessMemory().then(stats => {
+function getChromeMemory(options) {
+  return chromeMemory.getMemory(options.chromeProcessArgs).then(stats => {
     if (!stats.length) {
       // console.log('\nWARNING: No Chrome process found with the following arguments:')
       // console.log(chromeArguments);
@@ -197,7 +112,7 @@ function getChromeMemory(units) {
 
     // Only return the first process
     return {
-      chromeTotal: convertUnits(stats[0].memory, units)
+      chromeTotal: convertUnits(stats[0].memory, options.units)
     };
   });
 }
